+++ classes/class.ilTestArchiveCreatorHTML.php
@@ -34,8 +34,8 @@ class ilTestArchiveCreatorHTML
 		$this->tpl = new ilTestArchiveCreatorTemplate($this->plugin->getDirectory(). "/templates/tpl.content_page.html", true, true);
 		$GLOBALS['tpl'] = $this->tpl;
 
-        ilMathJax::getInstance()->init(ilMathJax::PURPOSE_EXPORT)
-                 ->setRendering(ilMathJax::RENDER_SVG_AS_XML_EMBED);
+        // render all MathJax at once in buildContent at the end
+        ilMathJax::getInstance()->init(ilMathJax::PURPOSE_DEFERRED_PDF);
     }
 
     /**
@@ -65,8 +65,24 @@ class ilTestArchiveCreatorHTML
         $tpl = new ilTestArchiveCreatorTemplate($this->plugin->getDirectory(). "/templates/tpl.content_page.html", true, true);
         $tpl->getDataFrom($this->tpl);
 
+
+        // Inclusion of MathJax script to the template is needed for STACK questions
+        // if server-side rendering is not enabled for browser
+
         if ($for_pdf) {
             $tpl->removeMediaPlayer();
+            $content = ilMathJax::getInstance()
+                                  ->init(ilMathJax::PURPOSE_PDF)
+                                  ->setRendering(ilMathJax::RENDER_SVG_AS_XML_EMBED)
+                                  ->includeMathJax($tpl)
+                                  ->insertLatexImages($content);
+        }
+        else {
+            $content = ilMathJax::getInstance()
+                                ->init(ilMathJax::PURPOSE_EXPORT)
+                                ->setRendering(ilMathJax::RENDER_SVG_AS_XML_EMBED)
+                                ->includeMathJax($tpl)
+                                ->insertLatexImages($content);
         }
 
         $tpl->addCss(ilUtil::getStyleSheetLocation('output', 'test_javascript.css', 'Modules/TestQuestionPool'), 'all');