+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## 1.6.1 for ILIAS 8 (upcoming)
 - fixed 0040794: Description regarding sample solutions is not precise
 - fixed 0040728: Text Subset Question with problematic display of best solution in html
+- Improved MathJax handling. MathJax settings for server-side rendering are respected separately: 
+  - 'Use for HTML Export' for the HTML files in the archive 
+  - 'Use for PDF Generation' for the optional PDF files. 
+  - 'Use for Browser' must be activated if TeX in STACK questions should be rendered server-side. 
+  - If server-side rendering is not enabled, then the Script URL client-side rendering is added to the HTML files in the archive.
 
 ## 1.6.0 for ILIAS 8 (2024-03-05)
 - added config option to include local assets (Media, CSS and JS File) in the archive